Notify user if no routes found and settings changed
